PyDispatch/PyDispatch_Desktop_Basic/requirements.txt
customtkinter
mysql-connector-python
Pillow
packaging
